Updating module to be less specific to chassism1000e
