+++ lib/puppet/type/cmc_fw_update.rb
@@ -1,4 +1,4 @@
-Puppet::Type.newtype(:chassism1000e_fw_update) do
+Puppet::Type.newtype(:cmc_fw_update) do
   desc "racadm util to update firmware"
 
   apply_to_device
