+++ lib/puppet/util/network_device/cmc.rb
@@ -1,0 +1,3 @@
+module Puppet::Util::NetworkDevice::Cmc
+
+end
