+++ metadata.json
@@ -1,8 +1,8 @@
 {
-  "name": "dell-chassism1000e",
+  "name": "dell-cmc",
   "version": "0.1.0",
   "author": "Logan M Kimmel",
-  "summary": "Custom puppet device that manages the Dell m1000e chassis",
+  "summary": "Custom puppet device that manages generic Dell chassis",
   "license": "Apache 2.0",
   "source": "git",
   "project_page": null,
