+++ spec/classes/init_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
-describe 'chassism1000e' do
+describe 'cmc' do
 
   context 'with defaults for all parameters' do
-    it { should contain_class('chassism1000e') }
+    it { should contain_class('cmc') }
   end
 end
