+++ templates/config.erb
@@ -1,3 +1,3 @@
 [<%= name %>]
-type chassism1000e
+type cmc
 url ssh://<%= username %>:<%= password %>@<%= url %>:<%= port %>
